Refactor simulation cluster management and improve gRPC handling
Abstracted Docker-based simulation cluster management into a new `SimulationClusterManager` class. Replaced ad-hoc gRPC stub creation with a reusable `GrpcStubManager` for better resource handling. Miscellaneous improvements include added dynamic worker count, logging enhancements, and inclusion of the `psutil` library for CPU-based scaling logic.
