+++ pyproject.toml
@@ -8,6 +8,7 @@ dependencies = [
     "grpcio==1.69.0",
     "grpcio-tools==1.69.0",
     "numpy==2.2.1",
+    "psutil>=7.0.0",
     "pydantic==2.10.4",
     "scipy==1.15.2",
     "tomli==2.2.1",
