+++ src/logger/logging.conf
@@ -41,5 +41,6 @@ datefmt=%Y-%m-%d %H:%M:%S
 class=logging.Formatter
 
 [formatter_simpleFormatter]
-format=%(module)-30s %(lineno)-4d - %(levelname)-8s - %(message)s
+format=%(asctime)s.%(msecs)03d - %(module)-30s %(lineno)-4d - %(levelname)-8s - %(message)s
+datefmt=%Y-%m-%d %H:%M:%S
 class=logging.Formatter
