+++ src/services/simulation_service/core/service/__init__.py
@@ -1,6 +1,8 @@
+from services.simulation_service.core.service.simulation_cluster_manager import (
+    simulation_cluster_context_manager,
+)
 from services.simulation_service.core.service.simulation_service import (
     SimulationService,
-    simulation_cluster_context_manager,
 )
 
 __all__ = ["SimulationService", "simulation_cluster_context_manager"]
