+++ src/services/simulation_service/core/service/simulation_service.py
@@ -1,17 +1,11 @@
-import os
-import subprocess
-from contextlib import contextmanager
 from pathlib import Path
-from typing import Any, Generator, Sequence
+from typing import Any, Sequence
 
 import grpc
 
-from logger import get_logger, log_docker_logs
+from logger import get_logger
 from services.simulation_service.core.infrastructure.generated import (
     simulation_messaging_pb2 as sm,
-)
-from services.simulation_service.core.infrastructure.generated import (
-    simulation_messaging_pb2_grpc as sm_grpc,
 )
 from services.simulation_service.core.models import (
     SimulationCase,
@@ -20,139 +14,15 @@ from services.simulation_service.core.models import (
     SimulationServiceResponse,
     WellManagementServiceResult,
 )
+from services.simulation_service.core.service.grpc_stub_manager import GrpcStubManager
 from services.simulation_service.core.utils.converters import json_to_str, str_to_json
 
 logger = get_logger(__name__)
-
-channel_options = [
-    ("grpc.max_send_message_length", 100 * 1024 * 1024),  # 100MB
-    ("grpc.max_receive_message_length", 100 * 1024 * 1024),  # 100MB
-]
-
-
-@contextmanager
-def core_directory() -> Generator[None, None, None]:
-    """
-    A context manager to temporarily switch to the core directory and restore the original directory upon exit.
-
-    Yields:
-        None
-    """
-    current_file = Path(__file__)
-    core_dir = current_file.parent.parent
-    logger.debug("Attempting to switch to core directory: %s", core_dir)
-
-    if not core_dir.exists():
-        logger.error("Core directory not found at %s", core_dir)
-        raise FileNotFoundError(f"Core directory not found at {core_dir}")
-
-    original_dir = os.getcwd()
-    logger.debug("Original working directory: %s", original_dir)
-
-    try:
-        os.chdir(core_dir)
-        logger.debug("Switched to core directory: %s", os.getcwd())
-        yield
-    finally:
-        os.chdir(original_dir)
-        logger.debug("Reverted to original working directory: %s", original_dir)
 
 
 class SimulationService:
     _SERVER_HOST = "localhost"
     _SERVER_PORT = 50051
-    _WORKER_COUNT = 3
-
-    @staticmethod
-    def build_simulation_cluster_images() -> None:
-        logger.info("Building simulation cluster images...")
-        with core_directory():
-            try:
-                result = subprocess.run(
-                    ["docker", "compose", "build", "--progress", "plain", "--force-rm"],
-                    check=True,
-                    capture_output=True,
-                    text=True,
-                )
-                logger.debug("Docker build output:\n%s", result.stdout)
-            except subprocess.CalledProcessError as e:
-                logger.error("Error building simulation cluster images:\n%s", e.stderr)
-                raise
-
-    @staticmethod
-    def start_simulation_cluster() -> None:
-        """
-        Start the Docker-based simulation cluster with the specified worker count.
-        """
-        logger.info(
-            "Attempting to start simulation cluster on host %s, port %d...",
-            SimulationService._SERVER_HOST,
-            SimulationService._SERVER_PORT,
-        )
-
-        # Clean up dangling images before starting cluster.
-        # Note: dangling images are images with both <none> as their repository and tag
-        try:
-            logger.info("Pruning dangling Docker images...")
-            result = subprocess.run(
-                ["docker", "image", "prune", "-f"],
-                check=True,
-                capture_output=True,
-                text=True,
-            )
-        except Exception as e:
-            logger.warning("Failed to prune Docker images: %s", str(e))
-            raise
-
-        with core_directory():
-            try:
-                result = subprocess.run(
-                    [
-                        "docker",
-                        "compose",
-                        "up",
-                        "-d",
-                        "--scale",
-                        f"worker={SimulationService._WORKER_COUNT}",
-                    ],
-                    check=True,
-                    capture_output=True,
-                    text=True,
-                )
-                logger.debug("Docker compose output:\n%s", result.stdout)
-                logger.info("Simulation cluster successfully started.")
-                log_docker_logs(logger)
-            except Exception as e:
-                logger.error(
-                    "Error starting the simulation cluster: %s",
-                    getattr(e, "stderr", str(e)),
-                )
-                raise
-
-            except subprocess.CalledProcessError as e:
-                logger.error("Error starting the simulation cluster:\n%s", e.stderr)
-                raise
-
-    @staticmethod
-    def shutdown_simulation_cluster() -> None:
-        """
-        Shutdown the Docker-based simulation cluster.
-        """
-        logger.info("Attempting to shut down the simulation cluster...")
-        with core_directory():
-            try:
-                subprocess.run(
-                    ["docker", "compose", "down"],
-                    check=True,
-                    capture_output=True,
-                    text=True,
-                )
-                logger.info("Simulation cluster successfully shut down.")
-            except subprocess.CalledProcessError as e:
-                logger.error(
-                    "Error shutting down the simulation cluster:\n%s", e.stderr
-                )
-                raise
 
     @staticmethod
     def process_request(request_dict: dict[str, Any]) -> SimulationServiceResponse:
@@ -169,11 +39,6 @@ class SimulationService:
         request = SimulationServiceRequest(**request_dict)
 
         simulation_cases_request = request.simulation_cases
-
-        logger.info(
-            "Processing %d simulation cases on the cluster...",
-            len(simulation_cases_request),
-        )
 
         simulation_cases_response = SimulationService._perform_simulations_on_cluster(
             simulation_cases_request
@@ -213,15 +78,10 @@ class SimulationService:
         Args:
             simulation_model_archive (bytes): The archive content as bytes.
         """
-
-        grpc_target = (
-            f"{SimulationService._SERVER_HOST}:{SimulationService._SERVER_PORT}"
-        )
-        logger.info("Establishing gRPC connection to %s...", grpc_target)
-
-        with grpc.insecure_channel(grpc_target, options=channel_options) as channel:
-            stub = sm_grpc.SimulationMessagingStub(channel)
-
+        with GrpcStubManager.get_stub(
+            SimulationService._SERVER_HOST,
+            SimulationService._SERVER_PORT,
+        ) as stub:
             try:
                 logger.info("Sending simulation model archive to the cluster...")
                 cluster_response = stub.TransferSimulationModel(
@@ -266,15 +126,12 @@ class SimulationService:
         Returns:
             Sequence[SimulationCase]: The processed simulation cases.
         """
+        logger.info("Processing %d simulation cases on the cluster...", len(cases))
 
-        grpc_target = (
-            f"{SimulationService._SERVER_HOST}:{SimulationService._SERVER_PORT}"
-        )
-        logger.info("Connecting to simulation cluster at %s...", grpc_target)
-
-        with grpc.insecure_channel(grpc_target, options=channel_options) as channel:
-            stub = sm_grpc.SimulationMessagingStub(channel)
-
+        with GrpcStubManager.get_stub(
+            SimulationService._SERVER_HOST,
+            SimulationService._SERVER_PORT,
+        ) as stub:
             logger.info("Sending simulation cases to the cluster...")
             simulations_inputs = [SimulationService._to_grpc(case) for case in cases]
             simulations_request = sm.Simulations(simulations=simulations_inputs)
@@ -325,18 +182,3 @@ class SimulationService:
             results=SimulationResults(**str_to_json(simulation.result.result)),
             control_vector=str_to_json(simulation.control_vector.content),
         )
-
-
-@contextmanager
-def simulation_cluster_context_manager():
-    """
-    Context manager for managing the simulation cluster lifecycle.
-    """
-    logger.info("Entering simulation cluster context.")
-    SimulationService.build_simulation_cluster_images()
-    SimulationService.start_simulation_cluster()
-    try:
-        yield
-    finally:
-        SimulationService.shutdown_simulation_cluster()
-        logger.info("Exited simulation cluster context.")
